+++ CHANGELOG.md
@@ -88,6 +88,7 @@ Date format: (YYYY-MM-DD)
   * Debug: Added the shopkeeper mob behavior tick period to the output of the 'check' command.
   * Metrics: The value of the new 'mob-behavior-tick-period' setting is captured by the features chart.
 * Performance: The rate at which the shopkeeper mob AI activations are updated for online players has been reduced from every 20 ticks to every 30 ticks. This seems sufficient to fluently activate nearby shopkeeper mobs even for players that fly around in creative mode.
+* We immediately activate the shopkeeper mob AI in nearby chunks now whenever a player joins the server or teleports.
 
 API:  
 * PlayerCreatePlayerShopkeeperEvent and PlayerShopkeeperHireEvent: The meaning of the max shops limit has changed. A value of 0 or less no longer indicates 'no limit'.